Fix for rockchip-otp to use legacy fixed of cells binding
The rockchip-otp driver uses the legacy fixed NVMEM cells directly under
the NVMEM device subnode.  This has been deprecated by the NVMEM subsystem.

Legacy drivers need to set add_legacy_fixed_of_cells to utilize the legacy
bindings.  Add the setting of this flag to rockchip-otp to allow it to
use this.  Without this fix none of the cells defined by OF get loaded.