arm64: dts: rockchip: rk3588-evb1: Couple CPU regulators
The RK3588 CPUs have two supply inputs: one supply for the logic and one
for the memory interface. On many platforms both supplies are handled by
the same regulator.

Boards, which have separate regulators for each supply need them coupled
together. This is necessary when cpufreq support is added to avoid crashes.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>