Fix/hack for VOP2 low performance
Remove drm_atomic_helper_dirtyfb as the .dirty handler.  As near as I can
tell this was used to force flushes on some other rockchip/vop variant.

However I don't believe this is needed for the rk3588.  It forced the
Xorg calls to DRM_IOCTL_MODE_DIRTYFB to block until vsync. However we
are directly operating on the front buffer so there is no benefit
(that I know of) to waiting for this.  Removing this improved my
llvmpipe glxgears test case from ~30 fps to ~52 fps.