Fix incorrect YUV420 override
Use logic from downstream rockchip driver in place of upstream.  The
upstream seems to have a botched attempt at simplifying the downstream
logic for when the tmdsclock is greater than the max_tmds_clock.  This
resulted in it reducing the color format to YUV420 even though the
tmdsclock was not over the max.