clk: rockchip: rk3588: mark hclk_nvm critical
HCLK_NVM is used for SFC, eMMC, SDIO and SDMMC. The dependency for
SFC, eMMC and SDIO is properly described, but SDMMC is missing the
information that it needs HCLK_NVM. Since HCLK_SD clock is sourced
from the firmware, adding this dependency is not trivially possible.

Considering, that almost any board will use one of the above interfaces,
let's just mark hclk_nvm explicitly as critical independently of the
linked clock.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>