Hack to move kgdb early in start_kernel
Quick/dirty hack to gain ability to set breakpoints earlier in boot.

Note: this was done without any particular understanding of the code at
hand.  Subjectively it seems to work, but there are likely problems with
it.  Almost certainly it is limited to only certain platforms/architectures.

"Works for me"