clk: rockchip: rk3588: drop unused code
All clocks are registered early using CLK_OF_DECLARE(), which marks
the DT node as processed. For the processed DT node the probe routine
is never called. Thus this whole code is never executed.

Note, that a follow-up patch plans to use late registered clocks, which
is done using a platform device. to ease reviewing it makes sense to first
of all remove this.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>