WIP: Hacky fix to clk-rk3588 GATE_LINK support
We need a driver on the GATE_LINK devices.  The existing code results
in dev.device = NULL.  The driver needs to provide the pm ops pointers
to pm_clk_suspend/pm_clk_resume which in turn will do the
enables/disables on all of the clocks on the list.

Manual binding via device_bind_driver() is almost certainly the wrong
long term way to do this.  Also should consider whether a separate
driver is in order or if putting the pm ops on the main clk-rk3588
as was done here is appropriate.